+++ precompiles/bank/bank.go
@@ -20,13 +20,13 @@ const (
 	PrecompileAddress string = "0x0000000000000000000000000000000000000804"
 
 	// GasBalanceOf defines the gas cost for a single ERC-20 balanceOf query
-	GasBalanceOf uint64 = 100 // TODO: get actual estimated gas cost
+	GasBalanceOf = 2_851
 
 	// GasTotalSupply defines the gas cost for a single ERC-20 totalSupply query
-	GasTotalSupply uint64 = 100 // TODO: get actual estimated gas cost
+	GasTotalSupply = 2_477
 
 	// GasSupplyOf defines the gas cost for a single ERC-20 supplyOf query, taken from totalSupply of ERC20
-	GasSupplyOf uint64 = 2_477
+	GasSupplyOf = 2_477
 )
 
 var _ vm.PrecompiledContract = &Precompile{}